Improve intermission frontend
