+++ frontend/src/pages/TournamentBracketPage.jsx
@@ -1,7 +1,7 @@
 import { useState, useEffect, createContext } from "react"
-import { BASEURL } from "../utils/globalUtils"
 import { useParams, useNavigate } from "react-router-dom"
 import { getTournament } from "../api"
+import { MatchupTile } from "../components/TournamentsBracket/MatchupTile"
 // import { createNextRoundArray } from "../components/TournamentsBracket/NextRoundBracket"
 import IntermissionDisplay from "../components/TournamentsBracket/IntermissionDisplay"
 import Navbar from "../components/Navbar/Navbar"
@@ -52,100 +52,6 @@ function getTotalNumberOfMatchupsBasedOnRound(roundNumber) {
 	const remainingTeamsInTournamentBasedOnRound = 16 / Math.pow(2, roundNumber)
 	return remainingTeamsInTournamentBasedOnRound
 }
-async function advanceTeam(accountId, tournamentId, setTournamentInformation) {
-	const response = await fetch(`${BASEURL}/tournaments/team/advance/`, {
-		method: "PATCH",
-		headers: {
-			"Content-Type": "application/json",
-		},
-		body: JSON.stringify({ accountId: accountId, tournamentId: tournamentId }),
-	})
-	const updatedTournamentInformation = await response.json()
-	setTournamentInformation(updatedTournamentInformation)
-	return updatedTournamentInformation
-}
-function AdvanceButton({ accountId, setTournamentInformation }) {
-	const { id } = useParams()
-	return (
-		<div className="advance-div">
-			<button
-				className="advance-btn"
-				onClick={(event) => {
-					advanceTeam(accountId, id, setTournamentInformation)
-					event.stopPropagation()
-				}}
-			>
-				Advance
-			</button>
-		</div>
-	)
-}
-
-function MatchupTile({
-	matchup,
-	tournamentInformation,
-	round,
-	setTournamentInformation,
-}) {
-	const navigate = useNavigate()
-	let team2 = matchup.team2
-
-	if (team2 == null) {
-		team2 = (
-			<div className="empty-team">
-				<h2>Match TBD</h2>
-			</div>
-		)
-	} else {
-		team2 = (
-			<div
-				className="team 2"
-				onClick={() => navigate(`/teams/${matchup.team2.accountId}`)}
-			>
-				{!tournamentInformation.participantsAdvancedToNextRound[
-					matchup.team2.accountId
-				] &&
-					round == tournamentInformation.currentRound &&
-					round != 4 && !tournamentInformation.participantsAdvancedToNextRound[matchup.team1.accountId] && (
-						<AdvanceButton
-							accountId={matchup.team2.accountId}
-							setTournamentInformation={setTournamentInformation}
-						/>
-					)}
-				<h2>{matchup.team2.name}</h2>
-			</div>
-		)
-	}
-	return (
-		<div className="matchup-tile">
-			<div className="teams">
-				<div
-					className="team 1"
-					onClick={() => navigate(`/teams/${matchup.team1.accountId}`)}
-				>
-					{!tournamentInformation.participantsAdvancedToNextRound[
-						matchup.team1.accountId
-					] &&
-						round == tournamentInformation.currentRound &&
-						round != 4 &&
-						!tournamentInformation.participantsAdvancedToNextRound[
-							matchup.team2.accountId
-						] && (
-							<AdvanceButton
-								accountId={matchup.team1.accountId}
-								setTournamentInformation={setTournamentInformation}
-							/>
-						)}
-					<h2>{matchup.team1.name}</h2>
-				</div>
-				<div className="versus">
-					<h2>VS</h2>
-				</div>
-				{team2}
-			</div>
-		</div>
-	)
-}
 
 async function loadTournamentInformation(
 	setTournamentInformation,
@@ -153,6 +59,7 @@ async function loadTournamentInformation(
 	setDisplayedMatchups
 ) {
 	const tournamentInformation = await getTournament(id)
+    console.log(tournamentInformation)
 	setTournamentInformation(tournamentInformation)
 	if (tournamentInformation.isActive == true) {
 		setDisplayedMatchups(tournamentInformation.rounds.round1)
@@ -171,13 +78,11 @@ export function BracketPage() {
 		loadTournamentInformation(setTournamentInformation, id, setDisplayedMatchups)
 	}, [isLoading])
 
-	if (isLoading && displayedMatchups.length == 0) {
-		return <TournamentLoading tournamentInformation={tournamentInformation} />
-	} else if (!isLoading && displayedMatchups.length == 0 && rounds == 1) {
+	if (!isLoading && displayedMatchups.length == 0 && rounds == 1) {
 		return (
-			<TournamentContext.Provider value={{ isLoading, setIsLoading }}>
-				<IntermissionDisplay tournamentInformation={tournamentInformation} />
-			</TournamentContext.Provider>
+			// <TournamentContext.Provider value={{ setIsLoading, tournamentInformation }}>
+				<IntermissionDisplay setIsLoading={setIsLoading} tournamentInformation={tournamentInformation}/>
+			// </TournamentContext.Provider>
 		)
 	}
